Refs #35948 - Updated routing
